Improve client side telemetry
